refactor: 파일명 변경 및 함수 분리
- helper 파일명을 separatorHelpers로 변경
- helper파일에 있던 정규식 관련 함수는 regexHelpers 파일로 분리하여 관리
