+++ src/utils/regexHelpers.js
@@ -1,0 +1,3 @@
+export const escapeRegExp = (text) => {
+  return text.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+};
